removed some duplication and complexity
