+++ forecast/options.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/aouyang1/go-forecaster/feature"
+	"github.com/aouyang1/go-forecaster/models"
 	"gonum.org/v1/gonum/dsp/window"
 )
 
@@ -88,6 +89,30 @@ func NewDefaultOptions() *Options {
 		Regularization:     []float64{0.0},
 		SeasonalityOptions: NewDefaultSeasonalityOptions(),
 	}
+}
+
+func (o *Options) NewLassoAutoOptions() *models.LassoAutoOptions {
+	lassoOpt := models.NewDefaultLassoAutoOptions()
+	if len(o.Regularization) > 0 {
+		lassoOpt.Lambdas = o.Regularization
+	} else {
+		o.Regularization = lassoOpt.Lambdas
+	}
+
+	lassoOpt.FitIntercept = false
+
+	lassoOpt.Iterations = o.Iterations
+	if o.Iterations == 0 {
+		lassoOpt.Iterations = models.DefaultIterations
+	}
+
+	lassoOpt.Tolerance = o.Tolerance
+	if o.Tolerance == 0 {
+		lassoOpt.Tolerance = models.DefaultTolerance
+	}
+
+	lassoOpt.Parallelization = o.Parallelization
+	return lassoOpt
 }
 
 const (
@@ -243,6 +268,77 @@ func NewWeeklySeasonalityConfig(orders int) SeasonalityConfig {
 type DSTOptions struct {
 	Enabled           bool     `json:"enabled"`
 	TimezoneLocations []string `json:"timezone_locations"`
+}
+
+func (d DSTOptions) AdjustTime(t []time.Time) []time.Time {
+	if !d.Enabled {
+		return t
+	}
+
+	offsets := loadLocationOffsets(d.TimezoneLocations)
+
+	newT := make([]time.Time, len(t))
+	for i := 0; i < len(t); i++ {
+		newT[i] = adjustTime(t[i], offsets)
+	}
+	return newT
+}
+
+func loadLocationOffsets(names []string) []locDstOffset {
+	var offsets []locDstOffset
+	for _, name := range names {
+		loc, err := time.LoadLocation(name)
+		if err != nil {
+			slog.Info("unable to load location, skipping", "location", name)
+			continue
+		}
+		offset := getLocationDSTOffset(loc)
+		offsets = append(offsets, locDstOffset{
+			loc:    loc,
+			offset: offset,
+		})
+	}
+	return offsets
+}
+
+func getLocationDSTOffset(loc *time.Location) int {
+	ctDec := time.Date(2024, 12, 1, 0, 0, 0, 0, loc)
+	ctJul := time.Date(2024, 6, 1, 0, 0, 0, 0, loc)
+	var offset int
+	if ctDec.IsDST() {
+		offset = deriveDSToffset(ctDec, ctJul)
+	} else {
+		offset = deriveDSToffset(ctJul, ctDec)
+	}
+	return offset
+}
+
+func deriveDSToffset(dstTime, stdTime time.Time) int {
+	_, dstOffset := dstTime.Zone()
+	_, stdOffset := stdTime.Zone()
+
+	return dstOffset - stdOffset
+}
+
+type locDstOffset struct {
+	loc    *time.Location
+	offset int
+}
+
+// adjustTime checks a time against all dst offsets ranges and adjusts it by checking if the time is
+// in a dst location offset range and finally averaging the cumulative offsets.
+func adjustTime(t time.Time, offsets []locDstOffset) time.Time {
+	var offsetSum int
+	for _, offset := range offsets {
+		locT := t.In(offset.loc)
+		if locT.IsDST() {
+			offsetSum += offset.offset
+		}
+	}
+	if len(offsets) == 0 {
+		return t
+	}
+	return t.Add(time.Duration(offsetSum/len(offsets)) * time.Second)
 }
 
 // MaxWeekendDurBuffer sets a limit of 1 day before or after the weekend begins at 00:00 Saturday
